requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
nltk==3.9.1
PyYAML==6.0.2
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
pyarrow==17.0.0
streamlit==1.39.0
plotly==5.24.1
